pyproject.toml
[project]
name = "blog-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-unfold>=0.71.0",
    "djlint>=1.36.4",
]
